Struct google_spanner1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the Spanner hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_spanner1 as spanner1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use spanner1::Spanner;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Spanner::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `instance_configs_get(...)`, `instance_configs_list(...)`, `instances_create(...)`, `instances_databases_create(...)`, `instances_databases_drop_database(...)`, `instances_databases_get(...)`, `instances_databases_get_ddl(...)`, `instances_databases_get_iam_policy(...)`, `instances_databases_list(...)`, `instances_databases_operations_cancel(...)`, `instances_databases_operations_delete(...)`, `instances_databases_operations_get(...)`, `instances_databases_operations_list(...)`, `instances_databases_sessions_begin_transaction(...)`, `instances_databases_sessions_commit(...)`, `instances_databases_sessions_create(...)`, `instances_databases_sessions_delete(...)`, `instances_databases_sessions_execute_batch_dml(...)`, `instances_databases_sessions_execute_sql(...)`, `instances_databases_sessions_execute_streaming_sql(...)`, `instances_databases_sessions_get(...)`, `instances_databases_sessions_list(...)`, `instances_databases_sessions_partition_query(...)`, `instances_databases_sessions_partition_read(...)`, `instances_databases_sessions_read(...)`, `instances_databases_sessions_rollback(...)`, `instances_databases_sessions_streaming_read(...)`, `instances_databases_set_iam_policy(...)`, `instances_databases_test_iam_permissions(...)`, `instances_databases_update_ddl(...)`, `instances_delete(...)`, `instances_get(...)`, `instances_get_iam_policy(...)`, `instances_list(...)`, `instances_operations_cancel(...)`, `instances_operations_delete(...)`, `instances_operations_get(...)`, `instances_operations_list(...)`, `instances_patch(...)`, `instances_set_iam_policy(...)` and `instances_test_iam_permissions(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

pub fn instances_databases_sessions_partition_query(
&self,
request: PartitionQueryRequest,
session: &str
) -> ProjectInstanceDatabaseSessionPartitionQueryCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a set of partition tokens that can be used to execute a query operation in parallel. Each of the returned partition tokens can be used by ExecuteStreamingSql to specify a subset of the query result to read. The same session and read-only transaction must be used by the PartitionQueryRequest used to create the partition tokens and the ExecuteSqlRequests that use the partition tokens.

Partition tokens become invalid when the session used to create them is deleted, is idle for too long, begins a new transaction, or becomes too old. When any of these happen, it is not possible to resume the query, and the whole operation must be restarted from the beginning.

Arguments

request - No description provided.

session - Required. The session used to create the partitions.

pub fn instances_databases_get(
&self,
name: &str
) -> ProjectInstanceDatabaseGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the state of a Cloud Spanner database.

Arguments

name - Required. The name of the requested database. Values are of the form projects/<project>/instances/<instance>/databases/<database>.

pub fn instances_databases_sessions_commit(
&self,
request: CommitRequest,
session: &str
) -> ProjectInstanceDatabaseSessionCommitCall<'a, C, A>

Create a builder to help you perform the following task:

Commits a transaction. The request includes the mutations to be applied to rows in the database.

Commit might return an ABORTED error. This can occur at any time; commonly, the cause is conflicts with concurrent transactions. However, it can also happen for a variety of other reasons. If Commit returns ABORTED, the caller should re-attempt the transaction from the beginning, re-using the same session.

Arguments

request - No description provided.

session - Required. The session in which the transaction to be committed is running.

pub fn instances_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectInstanceGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for an instance resource. Returns an empty policy if an instance exists but does not have a policy set.

Authorization requires spanner.instances.getIamPolicy on resource.

Arguments

request - No description provided.

resource - REQUIRED: The Cloud Spanner resource for which the policy is being retrieved. The format is projects/<project ID>/instances/<instance ID> for instance resources and projects/<project ID>/instances/<instance ID>/databases/<database ID> for database resources.

pub fn instances_databases_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectInstanceDatabaseGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for a database resource. Returns an empty policy if a database exists but does not have a policy set.

Authorization requires spanner.databases.getIamPolicy permission on resource.

Arguments

request - No description provided.

resource - REQUIRED: The Cloud Spanner resource for which the policy is being retrieved. The format is projects/<project ID>/instances/<instance ID> for instance resources and projects/<project ID>/instances/<instance ID>/databases/<database ID> for database resources.

pub fn instances_databases_sessions_read(
&self,
request: ReadRequest,
session: &str
) -> ProjectInstanceDatabaseSessionReadCall<'a, C, A>

Create a builder to help you perform the following task:

Reads rows from the database using key lookups and scans, as a simple key/value style alternative to ExecuteSql. This method cannot be used to return a result set larger than 10 MiB; if the read matches more data than that, the read fails with a FAILED_PRECONDITION error.

Reads inside read-write transactions might return ABORTED. If this occurs, the application should restart the transaction from the beginning. See Transaction for more details.

Larger result sets can be yielded in streaming fashion by calling StreamingRead instead.

Arguments

request - No description provided.

session - Required. The session in which the read should be performed.

pub fn instances_databases_sessions_execute_batch_dml(
&self,
request: ExecuteBatchDmlRequest,
session: &str
) -> ProjectInstanceDatabaseSessionExecuteBatchDmlCall<'a, C, A>

Create a builder to help you perform the following task:

Executes a batch of SQL DML statements. This method allows many statements to be run with lower latency than submitting them sequentially with ExecuteSql.

Statements are executed in order, sequentially. ExecuteBatchDmlResponse will contain a ResultSet for each DML statement that has successfully executed. If a statement fails, its error status will be returned as part of the ExecuteBatchDmlResponse. Execution will stop at the first failed statement; the remaining statements will not run.

ExecuteBatchDml is expected to return an OK status with a response even if there was an error while processing one of the DML statements. Clients must inspect response.status to determine if there were any errors while processing the request.

See more details in ExecuteBatchDmlRequest and ExecuteBatchDmlResponse.

Arguments

request - No description provided.

session - Required. The session in which the DML statements should be performed.

pub fn instances_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectInstanceSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on an instance resource. Replaces any existing policy.

Authorization requires spanner.instances.setIamPolicy on resource.

Arguments

request - No description provided.

resource - REQUIRED: The Cloud Spanner resource for which the policy is being set. The format is projects/<project ID>/instances/<instance ID> for instance resources and projects/<project ID>/instances/<instance ID>/databases/<database ID> for databases resources.

pub fn instances_databases_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectInstanceDatabaseTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns permissions that the caller has on the specified database resource.

Attempting this RPC on a non-existent Cloud Spanner database will result in a NOT_FOUND error if the user has spanner.databases.list permission on the containing Cloud Spanner instance. Otherwise returns an empty set of permissions.

Arguments

request - No description provided.

resource - REQUIRED: The Cloud Spanner resource for which permissions are being tested. The format is projects/<project ID>/instances/<instance ID> for instance resources and projects/<project ID>/instances/<instance ID>/databases/<database ID> for database resources.

pub fn instances_databases_list(
&self,
parent: &str
) -> ProjectInstanceDatabaseListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists Cloud Spanner databases.

Arguments

parent - Required. The instance whose databases should be listed. Values are of the form projects/<project>/instances/<instance>.

pub fn instances_databases_sessions_rollback(
&self,
request: RollbackRequest,
session: &str
) -> ProjectInstanceDatabaseSessionRollbackCall<'a, C, A>

Create a builder to help you perform the following task:

Rolls back a transaction, releasing any locks it holds. It is a good idea to call this for any transaction that includes one or more Read or ExecuteSql requests and ultimately decides not to commit.

Rollback returns OK if it successfully aborts the transaction, the transaction was already aborted, or the transaction is not found. Rollback never returns ABORTED.

Arguments

request - No description provided.

session - Required. The session in which the transaction to roll back is running.

pub fn instances_get(&self, name: &str) -> ProjectInstanceGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets information about a particular instance.

Arguments

name - Required. The name of the requested instance. Values are of the form projects/<project>/instances/<instance>.

pub fn instances_databases_operations_get(
&self,
name: &str
) -> ProjectInstanceDatabaseOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn instances_databases_sessions_streaming_read(
&self,
request: ReadRequest,
session: &str
) -> ProjectInstanceDatabaseSessionStreamingReadCall<'a, C, A>

Create a builder to help you perform the following task:

Like Read, except returns the result set as a stream. Unlike Read, there is no limit on the size of the returned result set. However, no individual row in the result set can exceed 100 MiB, and no column value can exceed 10 MiB.

Arguments

request - No description provided.

session - Required. The session in which the read should be performed.

pub fn instances_list(&self, parent: &str) -> ProjectInstanceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all instances in the given project.

Arguments

parent - Required. The name of the project for which a list of instances is requested. Values are of the form projects/<project>.

pub fn instances_databases_sessions_create(
&self,
request: CreateSessionRequest,
database: &str
) -> ProjectInstanceDatabaseSessionCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new session. A session can be used to perform transactions that read and/or modify data in a Cloud Spanner database. Sessions are meant to be reused for many consecutive transactions.

Sessions can only execute one transaction at a time. To execute multiple concurrent read-write/write-only transactions, create multiple sessions. Note that standalone reads and queries use a transaction internally, and count toward the one transaction limit.

Cloud Spanner limits the number of sessions that can exist at any given time; thus, it is a good idea to delete idle and/or unneeded sessions. Aside from explicit deletes, Cloud Spanner can delete sessions for which no operations are sent for more than an hour. If a session is deleted, requests to it return NOT_FOUND.

Idle sessions can be kept alive by sending a trivial SQL query periodically, e.g., "SELECT 1".

Arguments

request - No description provided.

database - Required. The database in which the new session is created.

pub fn instances_databases_create(
&self,
request: CreateDatabaseRequest,
parent: &str
) -> ProjectInstanceDatabaseCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new Cloud Spanner database and starts to prepare it for serving. The returned long-running operation will have a name of the format <database_name>/operations/<operation_id> and can be used to track preparation of the database. The metadata field type is CreateDatabaseMetadata. The response field type is Database, if successful.

Arguments

request - No description provided.

parent - Required. The name of the instance that will serve the new database. Values are of the form projects/<project>/instances/<instance>.

pub fn instances_databases_update_ddl(
&self,
request: UpdateDatabaseDdlRequest,
database: &str
) -> ProjectInstanceDatabaseUpdateDdlCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the schema of a Cloud Spanner database by creating/altering/dropping tables, columns, indexes, etc. The returned long-running operation will have a name of the format <database_name>/operations/<operation_id> and can be used to track execution of the schema change(s). The metadata field type is UpdateDatabaseDdlMetadata. The operation has no response.

Arguments

request - No description provided.

database - Required. The database to update.

pub fn instances_databases_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectInstanceDatabaseSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on a database resource. Replaces any existing policy.

Authorization requires spanner.databases.setIamPolicy permission on resource.

Arguments

request - No description provided.

resource - REQUIRED: The Cloud Spanner resource for which the policy is being set. The format is projects/<project ID>/instances/<instance ID> for instance resources and projects/<project ID>/instances/<instance ID>/databases/<database ID> for databases resources.

pub fn instances_patch(
&self,
request: UpdateInstanceRequest,
name: &str
) -> ProjectInstancePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an instance, and begins allocating or releasing resources as requested. The returned long-running operation can be used to track the progress of updating the instance. If the named instance does not exist, returns NOT_FOUND.

Immediately upon completion of this request:

For resource types for which a decrease in the instance's allocation has been requested, billing is based on the newly-requested level.

Until completion of the returned operation:

Cancelling the operation sets its metadata's cancel_time, and begins restoring resources to their pre-request values. The operation is guaranteed to succeed at undoing all resource changes, after which point it terminates with a CANCELLED status.

All other attempts to modify the instance are rejected.

Reading the instance via the API continues to give the pre-request resource levels.

Upon completion of the returned operation:

Billing begins for all successfully-allocated resources (some types may have lower than the requested levels).

All newly-reserved resources are available for serving the instance's tables.

The instance's new resource levels are readable via the API.

The returned long-running operation will have a name of the format <instance_name>/operations/<operation_id> and can be used to track the instance modification. The metadata field type is UpdateInstanceMetadata. The response field type is Instance, if successful.

Authorization requires spanner.instances.update permission on resource name.

Arguments

request - No description provided.

name - Required. A unique identifier for the instance, which cannot be changed after the instance is created. Values are of the form projects/<project>/instances/a-z*[a-z0-9]. The final segment of the name must be between 2 and 64 characters in length.

pub fn instances_databases_sessions_delete(
&self,
name: &str
) -> ProjectInstanceDatabaseSessionDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Ends a session, releasing server resources associated with it. This will asynchronously trigger cancellation of any operations that are running with this session.

Arguments

name - Required. The name of the session to delete.

pub fn instance_configs_get(
&self,
name: &str
) -> ProjectInstanceConfigGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets information about a particular instance configuration.

Arguments

name - Required. The name of the requested instance configuration. Values are of the form projects/<project>/instanceConfigs/<config>.

pub fn instances_databases_sessions_execute_sql(
&self,
request: ExecuteSqlRequest,
session: &str
) -> ProjectInstanceDatabaseSessionExecuteSqlCall<'a, C, A>

Create a builder to help you perform the following task:

Executes an SQL statement, returning all results in a single reply. This method cannot be used to return a result set larger than 10 MiB; if the query yields more data than that, the query fails with a FAILED_PRECONDITION error.

Operations inside read-write transactions might return ABORTED. If this occurs, the application should restart the transaction from the beginning. See Transaction for more details.

Larger result sets can be fetched in streaming fashion by calling ExecuteStreamingSql instead.

Arguments

request - No description provided.

session - Required. The session in which the SQL query should be performed.

pub fn instances_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectInstanceTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns permissions that the caller has on the specified instance resource.

Attempting this RPC on a non-existent Cloud Spanner instance resource will result in a NOT_FOUND error if the user has spanner.instances.list permission on the containing Google Cloud Project. Otherwise returns an empty set of permissions.

Arguments

request - No description provided.

resource - REQUIRED: The Cloud Spanner resource for which permissions are being tested. The format is projects/<project ID>/instances/<instance ID> for instance resources and projects/<project ID>/instances/<instance ID>/databases/<database ID> for database resources.

pub fn instances_databases_get_ddl(
&self,
database: &str
) -> ProjectInstanceDatabaseGetDdlCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the schema of a Cloud Spanner database as a list of formatted DDL statements. This method does not show pending schema updates, those may be queried using the Operations API.

Arguments

database - Required. The database whose schema we wish to get.

pub fn instances_databases_sessions_begin_transaction(
&self,
request: BeginTransactionRequest,
session: &str
) -> ProjectInstanceDatabaseSessionBeginTransactionCall<'a, C, A>

Create a builder to help you perform the following task:

Begins a new transaction. This step can often be skipped: Read, ExecuteSql and Commit can begin a new transaction as a side-effect.

Arguments

request - No description provided.

session - Required. The session in which the transaction runs.

pub fn instances_operations_list(
&self,
name: &str
) -> ProjectInstanceOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED.

NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn instance_configs_list(
&self,
parent: &str
) -> ProjectInstanceConfigListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the supported instance configurations for a given project.

Arguments

parent - Required. The name of the project for which a list of supported instance configurations is requested. Values are of the form projects/<project>.

pub fn instances_databases_sessions_partition_read(
&self,
request: PartitionReadRequest,
session: &str
) -> ProjectInstanceDatabaseSessionPartitionReadCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a set of partition tokens that can be used to execute a read operation in parallel. Each of the returned partition tokens can be used by StreamingRead to specify a subset of the read result to read. The same session and read-only transaction must be used by the PartitionReadRequest used to create the partition tokens and the ReadRequests that use the partition tokens. There are no ordering guarantees on rows returned among the returned partition tokens, or even within each individual StreamingRead call issued with a partition_token.

Partition tokens become invalid when the session used to create them is deleted, is idle for too long, begins a new transaction, or becomes too old. When any of these happen, it is not possible to resume the read, and the whole operation must be restarted from the beginning.

Arguments

request - No description provided.

session - Required. The session used to create the partitions.

pub fn instances_databases_sessions_execute_streaming_sql(
&self,
request: ExecuteSqlRequest,
session: &str
) -> ProjectInstanceDatabaseSessionExecuteStreamingSqlCall<'a, C, A>

Create a builder to help you perform the following task:

Like ExecuteSql, except returns the result set as a stream. Unlike ExecuteSql, there is no limit on the size of the returned result set. However, no individual row in the result set can exceed 100 MiB, and no column value can exceed 10 MiB.

Arguments

request - No description provided.

session - Required. The session in which the SQL query should be performed.

pub fn instances_operations_get(
&self,
name: &str
) -> ProjectInstanceOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn instances_databases_operations_cancel(
&self,
name: &str
) -> ProjectInstanceDatabaseOperationCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

name - The name of the operation resource to be cancelled.

pub fn instances_databases_sessions_list(
&self,
database: &str
) -> ProjectInstanceDatabaseSessionListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all sessions in a given database.

Arguments

database - Required. The database in which to list sessions.

pub fn instances_databases_operations_list(
&self,
name: &str
) -> ProjectInstanceDatabaseOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED.

NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn instances_create(
&self,
request: CreateInstanceRequest,
parent: &str
) -> ProjectInstanceCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an instance and begins preparing it to begin serving. The returned long-running operation can be used to track the progress of preparing the new instance. The instance name is assigned by the caller. If the named instance already exists, CreateInstance returns ALREADY_EXISTS.

Immediately upon completion of this request:

The instance is readable via the API, with all requested attributes but no allocated resources. Its state is CREATING.

Until completion of the returned operation:

Cancelling the operation renders the instance immediately unreadable via the API.

The instance can be deleted.

All other attempts to modify the instance are rejected.

Upon completion of the returned operation:

Billing for all successfully-allocated resources begins (some types may have lower than the requested levels).

Databases can be created in the instance.

The instance's allocated resource levels are readable via the API.

The instance's state becomes READY.

The returned long-running operation will have a name of the format <instance_name>/operations/<operation_id> and can be used to track creation of the instance. The metadata field type is CreateInstanceMetadata. The response field type is Instance, if successful.

Arguments

request - No description provided.

parent - Required. The name of the project in which to create the instance. Values are of the form projects/<project>.

pub fn instances_databases_drop_database(
&self,
database: &str
) -> ProjectInstanceDatabaseDropDatabaseCall<'a, C, A>

Create a builder to help you perform the following task:

Drops (aka deletes) a Cloud Spanner database.

Arguments

database - Required. The database to be dropped.

pub fn instances_operations_delete(
&self,
name: &str
) -> ProjectInstanceOperationDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

pub fn instances_databases_operations_delete(
&self,
name: &str
) -> ProjectInstanceDatabaseOperationDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

pub fn instances_delete(
&self,
name: &str
) -> ProjectInstanceDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an instance.

Immediately upon completion of the request:

Billing ceases for all of the instance's reserved resources.

Soon afterward:

The instance and all of its databases immediately and irrevocably disappear from the API. All data in the databases is permanently deleted.

Arguments

name - Required. The name of the instance to be deleted. Values are of the form projects/<project>/instances/<instance>

pub fn instances_operations_cancel(
&self,
name: &str
) -> ProjectInstanceOperationCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

name - The name of the operation resource to be cancelled.

pub fn instances_databases_sessions_get(
&self,
name: &str
) -> ProjectInstanceDatabaseSessionGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a session. Returns NOT_FOUND if the session does not exist. This is mainly useful for determining whether a session is still alive.

Arguments

name - Required. The name of the session to retrieve.